requirements.txt

fastapi ~=0.115.5
google-generativeai~=0.5.4
langchain-core ~=0.3.33
langchain-community ~=0.3.16
langchain-anthropic ~= 0.3.0
langchain-aws ~=0.2.14
langchain-chroma ~=0.2.3
langchain-google-genai ~=2.0.11
langchain-google-vertexai>=2.0.7
langchain-groq ~=0.2.1
langchain-ollama ~=0.2.3
langchain-openai ~=0.3.0
langfuse ~=2.60.5
langgraph ~=0.5.0
langgraph-checkpoint-mongodb ~=0.1.3
langgraph-checkpoint-postgres ~=2.0.13
langgraph-checkpoint-sqlite ~=2.0.1
langgraph-supervisor ~=0.0.27
langsmith ~=0.4.0
pydantic ~=2.10.1
pydantic-settings ~=2.6.1
python-dotenv ~=1.0.1
setuptools ~=75.6.0
uvicorn ~=0.32.1
python-jose ~=3.3.0
passlib ~=1.7.4
python-multipart ~=0.0.2